refactor: extract inline lists into feature components
- Extract inline task list logic from all-tasks page into reusable TasksList component
- Extract inline invitation error state into reusable InvitationFetchError component
