+++ features/invitation/index.ts
@@ -9,4 +9,5 @@ export { InvitationHeader } from './components/invitation-header';
 export { InvitationStats } from './components/invitation-stats';
 export { InvitationOverview } from './components/invitation-overview';
 export { InvitationEmptyState } from './components/invitation-empty-state';
+export { InvitationFetchError } from './components/invitation-fetch-error';
 export { InvitationForm } from './components/invitation-form';
