+++ features/tasks/components/index.ts
@@ -1,4 +1,5 @@
 export { TaskStatsCard } from './task-stats-card';
+export { TasksList } from './tasks-list';
 export { TaskCharts } from './task-charts';
 export { TaskAttachmentsSection } from './task-attachments-section';
 export { TaskAttachmentsUploader } from './task-attachments-uploader';
